Move `Signed` gadget from MULH opcode to `gadgets` submodule
