+++ ceno_zkvm/src/gadgets/signed_ext.rs
@@ -9,11 +9,16 @@ use crate::{
 use ff_ext::ExtensionField;
 use std::{marker::PhantomData, mem::MaybeUninit};
 
+/// Extract the most significant bit from an expression previously constrained
+/// to an 8- or 16-bit length.
+///
+/// Uses 1 `WitIn` value to store the bit, one `assert_bit` constraint, and one
+/// `u8` or `u16` table lookup.
 #[derive(Debug)]
 pub struct SignedExtendConfig<E> {
-    /// most significant bit
+    /// Most significant bit
     msb: WitIn,
-    /// number of bits contained in the value
+    /// Number of bits of the represented value
     n_bits: usize,
 
     _marker: PhantomData<E>,
